+++ src/main/java/com/epam/laboratory/app/service/TrainerServiceImpl.java
@@ -1,13 +1,16 @@
 package com.epam.laboratory.app.service;
 
-import com.epam.laboratory.app.aspect.Logging;
+import com.epam.laboratory.app.aspect.annotation.Logging;
+import com.epam.laboratory.app.domain.Trainee;
 import com.epam.laboratory.app.domain.Trainer;
+import com.epam.laboratory.app.exception.NoSuchEntityException;
 import com.epam.laboratory.app.repository.TrainerDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
+import java.util.Collection;
+import java.util.List;
 
 import static org.slf4j.event.Level.INFO;
 
@@ -15,30 +18,88 @@ import static org.slf4j.event.Level.INFO;
 @Transactional(rollbackFor = Exception.class)
 public class TrainerServiceImpl extends AbstractUserService<Trainer> implements TrainerService {
 
-    public TrainerServiceImpl(@Autowired TrainerDao trainerDao) {
-        super(trainerDao);
+    private final TraineeService traineeService;
+
+    @Autowired
+    public TrainerServiceImpl(TrainerDao trainerDao,
+                              AuthenticationService authenticationService,
+                              TraineeService traineeService) {
+        super(trainerDao, authenticationService);
+        this.traineeService = traineeService;
+    }
+
+    @Logging(INFO)
+    @Override
+    public Trainer updateByUsername(String username, Trainer entity) {
+        if (username == null) {
+            throw new IllegalArgumentException("Trainer username must not be null");
+        }
+        if (username.isBlank()) {
+            throw new IllegalArgumentException("Trainer username must not be blank");
+        }
+        if (entity == null) {
+            throw new IllegalArgumentException("Trainer must not be null");
+        }
+        var updatedTrainer = ((TrainerDao) dao).updateByUsername(username, entity);
+        updatedTrainer.getTrainees();
+        return updatedTrainer;
     }
 
     @Logging(INFO)
     @Transactional(readOnly = true)
     @Override
-    public Optional<Trainer> selectByUsername(String username) {
-        return ((TrainerDao) dao).findByUsername(username);
+    public Trainer selectByUsername(String username) {
+        if (username == null) {
+            throw new IllegalArgumentException("Trainer username must not be null");
+        }
+        if (username.isBlank()) {
+            throw new IllegalArgumentException("Trainer username must not be blank");
+        }
+        var optionalTrainer = ((TrainerDao) dao).findByUsername(username);
+        var trainer = optionalTrainer.orElseThrow(() ->
+                new NoSuchEntityException("Trainer with username " + username + " not found"));
+        trainer.getTrainees();
+        return trainer;
     }
 
     @Logging(INFO)
-    @Transactional(readOnly = true)
-    @Override
-    public boolean checkPasswordForUsername(String userName, String password) {
-        var trainerOptional = ((TrainerDao) dao).findByUsername(userName);
-        return trainerOptional.map(Trainer::getPassword)
-                .filter(password::equals)
-                .isPresent();
-    }
-
     @Override
     public void deleteByUsername(String username) {
+        if (username == null) {
+            throw new IllegalArgumentException("Trainer username must not be null");
+        }
+        if (username.isBlank()) {
+            throw new IllegalArgumentException("Trainer username must not be blank");
+        }
+        var isExists = authenticationService.checkExistsByUsername(username);
+        if (!isExists) {
+            throw new NoSuchEntityException("Trainer with username " + username + " not found");
+        }
         ((TrainerDao) dao).deleteByUsername(username);
     }
 
+    @Logging(INFO)
+    @Override
+    public void changeStatus(String username, boolean isActive) {
+        var isExists = authenticationService.checkExistsByUsername(username);
+        if (!isExists) {
+            throw new NoSuchEntityException("Trainer with username " + username + " not found");
+        }
+        ((TrainerDao) dao).changeStatusByUsername(username, isActive);
+    }
+
+    @Logging(INFO)
+    @Override
+    public Collection<Trainee> updateTrainees(String trainerUsername, Collection<Trainee> trainees) {
+        var trainer = selectByUsername(trainerUsername);
+        List<Trainee> traineesToSet = trainees.stream()
+                .map(Trainee::getUsername)
+                .map(traineeService::selectByUsername)
+                .toList();
+
+        trainer.getTrainees();
+        trainer.addTrainees(traineesToSet);
+        var updatedTrainer = update(trainer);
+        return updatedTrainer.getTrainees();
+    }
 }
